+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * deprecated constructing a `TreeBuilder` without passing root node information
  * renamed `FileLoaderLoadException` to `LoaderLoadException`
+ * deprecated using environment variables with `cannotBeEmpty()` if the value is validated with `validate()`
 
 4.1.0
 -----